Add "Remember Me" to the login form
Add a remember me token to the User model.
Add a remember me field to the login form.
Add methods to the auth_service to set and use the remember me token.
Add a design decision comment to auth_service to explain everything.
Add an @app.before_request method to check for remember me tokens.
(This is an ugly hack to deal with a circular import issue)

Minor changes:
Fix imports in read_views.