Add CSRF protection to all forms
Add CSRF methods to auth_svc.
Add CSRF tokens and token validation to the login and editor views.
Add CSRF tokens to the jinja templates.
Modify the new_session method to add a csrf token list.

Minor Changes:
Switch the "delete" function to a form to enable CSRF protection.
Rename the @app.before_request function to new_session().
Rewrite the design decision for the "Remember Me" tokens.
Add the |safe modifier to the jinja templates for the posts.